Provisioning: consolidate small integration test packages
Merge 7 small test packages into compatible larger ones to reduce
shared-server startup overhead. Each eliminated package saves one
Grafana server start (~7-15s).

Merges:
- jobs/flagdisabled, jobs/instanceauth, jobs/conflict -> jobs/
- foldermetadata/gitexport, git/foldermetadata_files, foldermetadata/fix -> foldermetadata/
- jobs/foldermetadata -> foldermetadata/

Key changes:
- foldermetadata/ switched from SharedEnv to SharedGitEnv to support
  git-backed tests; existing non-git tests use sharedHelper(t) returning
  *ProvisioningTestHelper, git tests use sharedGitHelper(t)
- conflict test fixed to use DeepCopy() instead of a second Get() call,
  removing the need for DisableControllers
- Name collisions in foldermetadata/ resolved by reusing existing helper
  functions via helper.ProvisioningTestHelper

Result: 19 packages -> 12 packages (7 fewer server starts).
Made-with: Cursor
