+++ pkg/tests/apis/provisioning/README.md
@@ -144,6 +144,37 @@ All shared test infrastructure lives in `common/testing.go`:
 | `RunGrafanaWithGitServer(t, opts...)` | Starts a standalone Grafana + git server for a single test |
 | `GrafanaOption` | Functional option type (`WithProvisioningFolderMetadata`, `WithRepositoryTypes`, etc.) |
 
+## Package Layout
+
+### SharedEnv packages (no git server)
+
+| Package | Env Options |
+| ------- | ----------- |
+| `provisioning/` (root) | `SecretsManagerEnableDBMigrations` |
+| `connection/` | none |
+| `foldermetadata/` | `WithProvisioningFolderMetadata` (SharedGitEnv — also hosts git-backed tests) |
+| `repository/` | `WithProvisioningFolderMetadata` |
+| `quota/` | none |
+| `jobs/` | none (GitHub mock in helper) |
+| `jobs/disabledfeatures` | `WithoutExportFeatureFlag` |
+| `enterprise/` | enterprise repo types (skipped in OSS) |
+
+### SharedGitEnv packages (git server required)
+
+| Package | Env Options |
+| ------- | ----------- |
+| `git/` | none |
+| `git/sourcepath_guard/` | none |
+| `git/foldermetadata_incremental/` | `WithProvisioningFolderMetadata` |
+| `foldermetadata/full/` | none |
+
+### Mixed-env package
+
+`foldermetadata/` uses `SharedGitEnv` with `WithProvisioningFolderMetadata` to support both
+local-filesystem tests and git-backed tests in a single package. The existing non-git tests
+use `sharedHelper(t)` (returns `*ProvisioningTestHelper`), while git-backed tests use
+`sharedGitHelper(t)` (returns `*GitTestHelper`).
+
 ## Migration Checklist for Other Packages
 
 1. Create `helpers_test.go` with `common.NewSharedEnv(...)` or `common.NewSharedGitEnv(...)` + helper func + `TestMain`
