+++ pkg/tests/apis/provisioning/foldermetadata/create_folder_git_test.go
@@ -1,4 +1,4 @@
-package foldermetadatafiles
+package foldermetadata
 
 import (
 	"context"
