+++ pkg/tests/apis/provisioning/foldermetadata/fixfoldermetadata_git_test.go
@@ -1,4 +1,4 @@
-package fix
+package foldermetadata
 
 import (
 	"context"
