--- pkg/tests/apis/provisioning/foldermetadata/gitexport/helper_test.go
@@ -1,18 +1,0 @@
-package gitexport
-
-import (
-	"testing"
-
-	"github.com/grafana/grafana/pkg/tests/apis/provisioning/common"
-)
-
-var env = common.NewSharedGitEnv(common.WithProvisioningFolderMetadata)
-
-func sharedGitHelper(t *testing.T) *common.GitTestHelper {
-	t.Helper()
-	return env.GetCleanHelper(t)
-}
-
-func TestMain(m *testing.M) {
-	env.RunTestMain(m)
-}
