+++ pkg/tests/apis/provisioning/jobs/conflict_test.go
@@ -1,4 +1,4 @@
-package conflict
+package jobs
 
 import (
 	"context"
@@ -40,8 +40,7 @@ func TestIntegrationProvisioning_JobConflict(t *testing.T) {
 
 	job, err := helper.Jobs.Resource.Get(ctx, createdJob.GetName(), metav1.GetOptions{})
 	require.NoError(t, err)
-	job2, err := helper.Jobs.Resource.Get(ctx, createdJob.GetName(), metav1.GetOptions{})
-	require.NoError(t, err)
+	job2 := job.DeepCopy()
 
 	client1Update := job.DeepCopy()
 	if client1Update.GetLabels() == nil {
