+++ pkg/tests/apis/provisioning/jobs/exportjob_auth_test.go
@@ -1,4 +1,4 @@
-package instanceauth
+package jobs
 
 import (
 	"context"
@@ -12,11 +12,11 @@ import (
 	"github.com/grafana/grafana/pkg/tests/apis/provisioning/common"
 )
 
-func TestIntegrationProvisioning_MigrateJobAuthorization(t *testing.T) {
+func TestIntegrationProvisioning_ExportJobAuthorization(t *testing.T) {
 	helper := sharedHelper(t)
 	ctx := context.Background()
 
-	const repo = "migrate-auth-test"
+	const repo = "export-auth-test"
 	testRepo := common.TestRepo{
 		Name:               repo,
 		Target:             "instance",
@@ -26,10 +26,10 @@ func TestIntegrationProvisioning_MigrateJobAuthorization(t *testing.T) {
 	}
 	helper.CreateRepo(t, testRepo)
 
-	t.Run("admin can create migrate job", func(t *testing.T) {
+	t.Run("admin can create export job", func(t *testing.T) {
 		body := common.AsJSON(provisioning.JobSpec{
-			Action:  provisioning.JobActionMigrate,
-			Migrate: &provisioning.MigrateJobOptions{},
+			Action: provisioning.JobActionPush,
+			Push:   &provisioning.ExportJobOptions{},
 		})
 
 		var statusCode int
@@ -42,16 +42,16 @@ func TestIntegrationProvisioning_MigrateJobAuthorization(t *testing.T) {
 			SetHeader("Content-Type", "application/json").
 			Do(ctx).StatusCode(&statusCode)
 
-		require.NoError(t, result.Error(), "admin should be able to create migrate job")
+		require.NoError(t, result.Error(), "admin should be able to create export job")
 		require.Equal(t, http.StatusAccepted, statusCode, "should return 202 Accepted")
 
 		helper.AwaitJobs(t, repo)
 	})
 
-	t.Run("editor cannot create migrate job on instance-scoped repo", func(t *testing.T) {
+	t.Run("editor cannot create export job on instance-scoped repo", func(t *testing.T) {
 		body := common.AsJSON(provisioning.JobSpec{
-			Action:  provisioning.JobActionMigrate,
-			Migrate: &provisioning.MigrateJobOptions{},
+			Action: provisioning.JobActionPush,
+			Push:   &provisioning.ExportJobOptions{},
 		})
 
 		var statusCode int
@@ -64,15 +64,15 @@ func TestIntegrationProvisioning_MigrateJobAuthorization(t *testing.T) {
 			SetHeader("Content-Type", "application/json").
 			Do(ctx).StatusCode(&statusCode)
 
-		require.Error(t, result.Error(), "editor should not be able to migrate on instance-scoped repo")
+		require.Error(t, result.Error(), "editor should not be able to export on instance-scoped repo")
 		require.Equal(t, http.StatusForbidden, statusCode, "should return 403 Forbidden")
 		require.True(t, apierrors.IsForbidden(result.Error()), "error should be forbidden")
 	})
 
-	t.Run("viewer cannot create migrate job", func(t *testing.T) {
+	t.Run("viewer cannot create export job", func(t *testing.T) {
 		body := common.AsJSON(provisioning.JobSpec{
-			Action:  provisioning.JobActionMigrate,
-			Migrate: &provisioning.MigrateJobOptions{},
+			Action: provisioning.JobActionPush,
+			Push:   &provisioning.ExportJobOptions{},
 		})
 
 		var statusCode int
@@ -85,7 +85,7 @@ func TestIntegrationProvisioning_MigrateJobAuthorization(t *testing.T) {
 			SetHeader("Content-Type", "application/json").
 			Do(ctx).StatusCode(&statusCode)
 
-		require.Error(t, result.Error(), "viewer should not be able to create migrate job")
+		require.Error(t, result.Error(), "viewer should not be able to create export job")
 		require.Equal(t, http.StatusForbidden, statusCode, "should return 403 Forbidden")
 		require.True(t, apierrors.IsForbidden(result.Error()), "error should be forbidden")
 	})
