+++ pkg/tests/apis/provisioning/jobs/flagdisabled_test.go
@@ -1,4 +1,4 @@
-package flagdisabled
+package jobs
 
 import (
 	"context"
