--- pkg/tests/apis/provisioning/jobs/foldermetadata/helpers_test.go
@@ -1,18 +1,0 @@
-package foldermetadata
-
-import (
-	"testing"
-
-	"github.com/grafana/grafana/pkg/tests/apis/provisioning/common"
-)
-
-var env = common.NewSharedEnv(common.WithProvisioningFolderMetadata)
-
-func sharedHelper(t *testing.T) *common.ProvisioningTestHelper {
-	t.Helper()
-	return env.GetCleanHelper(t)
-}
-
-func TestMain(m *testing.M) {
-	env.RunTestMain(m)
-}
